[merge]: Move DbRequest building functions to new module (#787)
